Driver: Update client/server model that is working
I recreated the client/server model that we had. Now the robots will
be the servers, who have an open socket listening for bluetooth
connections. When a computer wishes to update the software the bots
are running, they simply will run update_sim.py and pass the json file
with the new specification they want to use